+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # UNRELEASED
 
+-   Add `Unknown` to `EncryptionAlgorithm`, representing unsupported algorithms
+    coming from matrix-sdk-crypto's `EventEncryptionAlgorithm`.
+    ([#92](https://github.com/matrix-org/matrix-rust-sdk-crypto-wasm/pull/92))
+
 -   Add new methods `OlmMachine.{get,set}RoomSettings`.
     ([#95](https://github.com/matrix-org/matrix-rust-sdk-crypto-wasm/pull/95))